Various additions

+++ xen-install.sh
@@ -7,6 +7,8 @@ GIT_XEN_CO="RELEASE-4.7.0"
 # Here you can put for example:
 # XEN_COMPILE_OPTS="lock_profile=y perfc=y perfc_arrays=y"
 XEN_COMPILE_OPTS="lock_profile=y perfc=y perfc_arrays=y"
+# Should we automatically reboot after the installation ?
+REBOOT="no"
 # Xen git repository:
 GIT_XEN="git://xenbits.xen.org/xen"
 
@@ -16,8 +18,10 @@ CPUS=`cat /proc/cpuinfo | grep processor | wc -l`
 sudo apt-get update && sudo apt-get upgrade -y
 sudo apt-get install -y build-essential git python-dev bin86 bcc iasl uuid-dev libncurses5-dev libglib2.0-dev libpixman-1-dev libaio-dev libssl-dev libyajl-dev libc6-dev-i386 texinfo pandoc markdown transfig libnl-cli-3-dev libnl-3-200 libnl-cli-3-200 libnl-3-dev gettext libfdt-dev multiboot wget libpci-dev liblzma-dev
 
-git clone $GIT_XEN
-cd xen && git checkout $GIT_XEN_CO
+if [ ! -e xen ]; then
+	git clone $GIT_XEN
+	cd xen && git checkout $GIT_XEN_CO && cd -
+fi
 
 # hack to fix etherboot compile with gcc6
 #sed -i "s/IPXE_GIT_TAG := 9a93db3f0947484e30e753bbd61a10b17336e20e/IPXE_GIT_TAG := a4c4f72297bea6902001ce813aaf432bd49d382d/" tools/firmware/etherboot/Makefile
@@ -25,8 +29,8 @@ cd xen && git checkout $GIT_XEN_CO
 #echo "" > tools/firmware/etherboot/patches/series
 # End hack
 
-./configure
-$XEN_COMPILE_OPTS make -j$CPUS install
+cd xen && ./configure && cd -
+$XEN_COMPILE_OPTS make -j$CPUS -C xen install
 
 sudo ldconfig
 sudo update-rc.d xencommons defaults 19 18
@@ -43,6 +47,6 @@ sudo mv -f /etc/grub.d/20_linux_xen /etc/grub.d/09_linux_xen
 fi
 sudo update-grub
 
-cd ..
-
-reboot
+if [ "$REBOOT" == "yes" ]; then
+	reboot
+fi
